// src/app/models/tier.ts
export interface Tier {
    id?: number;
    codePrincipal: string;
    codeTier?: string;
    codeGS1?: string;
    codeRGP?: string;
    codeIdentificationDestinataire?: string;
    codeIdentificationExpediteur?: string;
    nomContactPrincipal: string;
    dateCreation: string;
    dateModification?: string;
    codePays?: string;
    // Ajoutez d'autres propriétés en fonction de votre besoin
  }
   